Rewrite partially Array implementation
I had designed the first version too quickly. As a result, it was both
not efficient as it would allocate small memory blocks for each items,
and incomplete. This commit add a simple useful forEach procedure,
resolve the memory allocation problem and changes the way resizing
works. That last point may not be perfect yet.